% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/take-snapshot.R
\name{take_snapshot}
\alias{take_snapshot}
\title{Take Snapshot of CRAN deadlines}
\usage{
take_snapshot()
}
\value{
A data.frame, one line per package with a deadline.
}
\description{
Take Snapshot of CRAN deadlines
}
\examples{
options(repos = "https://cran.r-rproject.org")
ts <- take_snapshot()
}
\author{
Code modified from Hadley Wickham
}
